requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.10.0=py37hc8dfbb8_1
appdirs=1.4.4=pyh9f0ad1d_0
argon2-cffi=20.1.0=py37h4abf009_2
astor=0.8.1=pyh9f0ad1d_0
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
black=20.8b1=py_1
bleach=3.2.1=pyh9f0ad1d_0
brotlipy=0.7.0=py37hb5d75c8_1001
c-ares=1.16.1=h516909a_3
ca-certificates=2020.11.8=ha878542_0
cached-property=1.5.1=py_0
certifi=2020.11.8=py37h89c1867_0
cffi=1.14.3=py37h00ebd2e_1
chardet=3.0.4=py37he5f6b98_1008
click=7.1.2=pyh9f0ad1d_0
cryptography=3.2.1=py37hc72a4ac_0
cycler=0.10.0=py_2
dataclasses=0.7=pyhb2cacf7_7
dbus=1.13.6=hfdff14a_1
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.2.9=he1b5a44_2
flask=1.1.2=pyh9f0ad1d_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.4=h7ca028e_0
gast=0.4.0=pyh9f0ad1d_0
gettext=0.19.8.1=hf34092f_1004
glib=2.66.2=h58526e2_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.33.2=py37haffed2e_2
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
h5py=2.10.0=nompi_py37hf7afa78_105
hdf5=1.10.6=nompi_h1022a3e_1110
icu=64.2=he1b5a44_1
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
ipdb=0.13.3=pyh9f0ad1d_0
ipykernel=5.3.4=py37h888b3d9_1
ipython=5.8.0=py37_1
ipython_genutils=0.2.0=py_1
isort=5.6.4=py_0
itsdangerous=1.1.0=py_0
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.17.0=py_0
jpeg=9d=h36c2ea0_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py37h89c1867_2
jupyterlab=2.2.9=py_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=1.2.0=py_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_0
kiwisolver=1.3.1=py37hc928c03_0
krb5=1.17.1=hfafb76e_3
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.35.1=hed1e6ac_0
libblas=3.9.0=2_openblas
libcblas=3.9.0=2_openblas
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=9.3.0=he4bcb1c_17
libgfortran5=9.3.0=he4bcb1c_17
libglib=2.66.2=hbe7bbb4_0
libgomp=9.3.0=h5dbcf3e_17
libiconv=1.16=h516909a_0
liblapack=3.9.0=2_openblas
libnghttp2=1.41.0=h8cfc5f6_2
libopenblas=0.3.12=pthreads_h4812303_1
libpng=1.6.37=h21135ba_2
libprotobuf=3.13.0.1=h8b12597_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.1.0=h4f3a223_6
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h36c2ea0_3
libxcb=1.13=h14c3975_1002
libxml2=2.9.10=hee79883_0
lz4-c=1.9.2=he1b5a44_3
markdown=3.3.3=pyh9f0ad1d_0
markupsafe=1.1.1=py37hb5d75c8_2
matplotlib=3.3.3=py37h89c1867_0
matplotlib-base=3.3.3=py37h4f6019d_0
mistune=0.8.4=py37h4abf009_1002
mypy_extensions=0.4.3=py37hc8dfbb8_2
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37h89c1867_3
nbformat=5.0.8=py_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.4.3=pyhd8ed1ab_0
notebook=6.1.5=py37h89c1867_0
numpy=1.19.4=py37h7e9df27_1
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1h=h516909a_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.4=py37h10a2094_0
pandoc=2.11.1.1=h36c2ea0_0
pandocfilters=1.4.2=py_1
pathspec=0.8.1=pyhd3deb0d_0
patsy=0.5.1=py_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.0.1=py37h63a5d19_0
pip=20.2.4=py_0
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt_toolkit=1.0.15=py_1
protobuf=3.13.0.1=py37h3340039_1
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py_1001
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.2=py_0
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py37hcca6a23_4
pyrsistent=0.17.3=py37h4abf009_1
pysocks=1.7.1=py37he5f6b98_2
python=3.7.8=h6f2ec95_1_cpython
python-dateutil=2.8.1=py_0
python-dotenv=0.15.0=pyhd8ed1ab_0
python_abi=3.7=1_cp37m
pytz=2020.4=pyhd8ed1ab_0
pyzmq=20.0.0=py37h5a562af_1
qt=5.9.7=h0c104cb_3
readline=8.0=he28a2e2_2
regex=2020.11.13=py37h4abf009_0
requests=2.25.0=pyhd3deb0d_0
scikit-learn=0.23.2=py37hddcf8d6_2
scipy=1.5.3=py37h14a347d_0
seaborn=0.11.0=ha770c72_1
seaborn-base=0.11.0=pyhd8ed1ab_1
send2trash=1.5.0=py_0
setuptools=49.6.0=py37he5f6b98_2
simplegeneric=0.8.1=py_1
sip=4.19.8=py37hf484d3e_0
six=1.15.0=pyh9f0ad1d_0
sqlite=3.33.0=h4cf870e_1
statsmodels=0.12.1=py37ha21ca33_1
tensorboard=1.14.0=py37_0
tensorflow=1.14.0=h4531e10_0
tensorflow-base=1.14.0=py37h4531e10_0
tensorflow-estimator=1.14.0=py37h5ca1d4c_0
termcolor=1.1.0=py_2
terminado=0.9.1=py37h89c1867_1
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hed695b0_1
toml=0.10.2=pyhd8ed1ab_0
tornado=6.1=py37h4abf009_0
traitlets=5.0.5=py_0
twilio=6.47.0=pyhd3deb0d_0
typed-ast=1.4.1=py37h4abf009_1
typing_extensions=3.7.4.3=py_0
urllib3=1.25.11=py_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.35.1=pyh9f0ad1d_0
wordcloud=1.8.1=py37h4abf009_0
wrapt=1.12.1=py37h8f50634_1
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xz=5.2.5=h516909a_1
zeromq=4.3.3=h58526e2_2
zipp=3.4.0=py_0
zlib=1.2.11=h516909a_1010
zstd=1.4.5=h6597ccf_2
